requirements.txt
opencv-python
numpy
tensorflow
flask
matplotlib
Pillow